# server/src/index.js
const express = require('express');
const cors = require('cors');
const { PORT } = require('./config/env');
const taskRoutes = require('./routes/task.routes');
const categoryRoutes = require('./routes/category.routes');

const swaggerUi = require('swagger-ui-express');
const swaggerJsdoc = require('swagger-jsdoc');

const swaggerOptions = {
  definition: {
    openapi: '3.0.0',
    info: {
      title: 'TaskFlow API',
      version: '1.0.0',
      description: 'API REST para gestionar tareas y categorías',
    },
    servers: [
      {
        url: 'https://taskflow-project-eight.vercel.app/api/v1',
      },
    ],
  },
  apis: ['./src/routes/*.js'],
};

const swaggerSpec = swaggerJsdoc(swaggerOptions);

const app = express();

app.use(cors());
app.use(express.json());
const swaggerDocument = require('../swagger.json');
app.use('/api/v1/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
app.use('/api/v1/tasks', taskRoutes);
app.use('/api/v1/categories', categoryRoutes);

app.get('/', (req, res) => {
  res.send('Servidor funcionando correctamente');
});

app.use((err, req, res, next) => {
  if (err.message === 'NOT_FOUND') {
    return res.status(404).json({ message: 'Recurso no encontrado' });
  }

  if (err.message === 'BAD_REQUEST') {
    return res.status(400).json({ message: 'Solicitud inválida' });
  }

  console.error(err);

  return res.status(500).json({ message: 'Error interno del servidor' });
});

if (require.main === module) {
  app.listen(PORT, () => {
    console.log(`Servidor escuchando en http://localhost:${PORT}`);
  });
}

module.exports = app;