# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TaskFlow API",
    "version": "1.0.0",
    "description": "API REST para gestionar tareas y categorías"
  },
  "paths": {
    "/categories": {
      "get": {
        "summary": "Obtener todas las categorías",
        "responses": {
          "200": {
            "description": "Lista de categorías"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva categoría",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoría creada correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
...
    "/tasks": {
      "get": {
        "summary": "Obtener todas las tareas",
        "responses": {
          "200": {
            "description": "Lista de tareas"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva tarea",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "texto": {
                    "type": "string"
                  },
                  "prioridad": {
                    "type": "string"
                  },
                  "categoria": {
                    "type": "string"
                  },
                  "fecha": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
...
  },
...
}